Refactor config out from components folder
